refactor(gui): relocate WFM demod lifecycle out of MainWindow.cpp (#3407 follow-up)
Pure code motion — no behaviour change. activateWFM / deactivateWFM /
reflectWfmButtons move from MainWindow.cpp into MainWindow_DigitalModes.cpp,
beside the analogous RADE/FreeDV/DAX demod-surface lifecycles carved out in
the #3351 Phase 1e split. They stay MainWindow:: members (header unchanged),
matching how activateRADE/deactivateRADE already live in that TU.

The RxApplet/VfoWidget wfmActivated connect handlers stay in MainWindow.cpp —
that's consistent with the RADE connects right next to them.

- MainWindow.cpp: −140 lines; drops the now-unused WfmDemodulator/
  WfmDeviceDialog/WfmSettings includes (m_wfmDemod only needs the existing
  forward-decl in MainWindow.h).
- MainWindow_DigitalModes.cpp: +147; carries the three WFM headers + RxApplet.h
  explicitly (Qt 6.4.2 floor — no reliance on transitive includes).

Builds clean; full suite (34 tests incl. wfm_dsp_test) green.
